+++ processor/k8seventgenerationprocessor/log_constructor.go
@@ -17,6 +17,7 @@ package k8seventgenerationprocessor
 import (
 	"os"
 
+	"github.com/solarwinds/solarwinds-otel-collector-contrib/processor/k8seventgenerationprocessor/internal/manifests"
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/plog"
 	conventions "go.opentelemetry.io/otel/semconv/v1.6.1"
@@ -36,6 +37,8 @@ const (
 	// Attributes for telemetry mapping
 	otelEntityId    = "otel.entity.id"
 	swK8sClusterUid = "sw.k8s.cluster.uid"
+	swK8sWorkloadIp = "sw.k8s.workload.ip"
+	serviceName     = "k8s.service.name"
 
 	// Attributes containing additional information about container
 	otelEntityAttributes = "otel.entity.attributes"
@@ -44,22 +47,22 @@ const (
 	k8sContainerSidecar  = "sw.k8s.container.sidecar"
 )
 
-// addContainersResourceLog adds a new ResourceLogs to the provided Logs structure
+// addContainerResourceLog adds a new ResourceLogs to the provided Logs structure
 // and sets required attributes on "resource" and "scopeLogs"
-func addContainersResourceLog(ld plog.Logs) plog.ResourceLogs {
+func addContainerResourceLog(ld plog.Logs, containersLogSlice plog.LogRecordSlice) {
 	rl := ld.ResourceLogs().AppendEmpty()
 	rl.Resource().Attributes().PutStr(k8sLogType, "entitystateevent")
 	sl := rl.ScopeLogs().AppendEmpty()
 	sl.Scope().Attributes().PutBool(otelEntityEventAsLog, true)
-	return rl
+	containersLogSlice.CopyTo(sl.LogRecords())
 }
 
 // transformManifestToContainerLogs returns a new plog.LogRecordSlice and appends
 // all LogRecords containing container information from the provided Manifest.
-func transformManifestToContainerLogs(m Manifest, t pcommon.Timestamp) plog.LogRecordSlice {
+func transformManifestToContainerLogs(m *manifests.PodManifest, t pcommon.Timestamp) plog.LogRecordSlice {
 	lrs := plog.NewLogRecordSlice()
 
-	containers := m.getContainers()
+	containers := m.GetContainers()
 	for _, c := range containers {
 		lr := lrs.AppendEmpty()
 		lr.SetObservedTimestamp(t)
@@ -70,14 +73,14 @@ func transformManifestToContainerLogs(m Manifest, t pcommon.Timestamp) plog.LogR
 }
 
 // addContainerAttributes sets attributes on the provided map for the given Metadata and Container.
-func addContainerAttributes(attrs pcommon.Map, md Metadata, c Container) {
+func addContainerAttributes(attrs pcommon.Map, md manifests.PodMetadata, c manifests.Container) {
 	// Ingestion attributes
 	attrs.PutStr(otelEntityEventType, entityState)
 	attrs.PutStr(swEntityType, k8sContainerEntityType)
 
 	// Telemetry mappings
 	tm := attrs.PutEmptyMap(otelEntityId)
-	tm.PutStr(string(conventions.K8SPodNameKey), md.PodName)
+	tm.PutStr(string(conventions.K8SPodNameKey), md.Name)
 	tm.PutStr(string(conventions.K8SNamespaceNameKey), md.Namespace)
 	tm.PutStr(string(conventions.K8SContainerNameKey), c.Name)
 	tm.PutStr(swK8sClusterUid, os.Getenv(clusterUidEnv))
@@ -88,4 +91,29 @@ func addContainerAttributes(attrs pcommon.Map, md Metadata, c Container) {
 	ea.PutStr(k8sContainerStatus, c.State)
 	ea.PutBool(k8sContainerInit, c.IsInitContainer)
 	ea.PutBool(k8sContainerSidecar, c.IsSidecarContainer)
+}
+
+// addServiceMappingsResourceLog adds a new ResourceLogs to the provided Logs structure
+// and sets required attributes on "resource" and "scopeLogs"
+func addServiceMappingsResourceLog(ld plog.Logs, serviceMappingsLogSlice plog.LogRecordSlice) {
+	rl := ld.ResourceLogs().AppendEmpty()
+	rl.Resource().Attributes().PutStr(k8sLogType, "serviceendpointsmapping")
+	lrs := rl.ScopeLogs().AppendEmpty().LogRecords()
+	serviceMappingsLogSlice.CopyTo(lrs)
+}
+
+func transformManifestToServiceMappingLogs(m manifests.ServiceMapping, t pcommon.Timestamp) plog.LogRecordSlice {
+	lrs := plog.NewLogRecordSlice()
+
+	for _, addr := range m.GetAddresses() {
+		lr := lrs.AppendEmpty()
+		lr.SetObservedTimestamp(t)
+		attrs := lr.Attributes()
+		attrs.PutStr(serviceName, m.GetServiceName())
+		attrs.PutStr(string(conventions.K8SNamespaceNameKey), m.GetNamespace())
+		attrs.PutStr(swK8sWorkloadIp, addr)
+		attrs.PutStr(swK8sClusterUid, os.Getenv(clusterUidEnv))
+	}
+
+	return lrs
 }